backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1

# FastAPI AI Chat service
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.5.1
httpx==0.27.2
pydantic==2.8.2
